+++ sdk/identity/identity/src/credentials/clientSecretCredential.browser.ts
@@ -66,7 +66,7 @@ export class ClientSecretCredential implements TokenCredential {
     options?: GetTokenOptions
   ): Promise<AccessToken | null> {
     const { span, updatedOptions: newOptions } = createSpan(
-      "ClientSecretCredential-getToken",
+      "ClientSecretCredential.getToken",
       options
     );